pyproject.toml
[project]
name = "kmm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "selenium==3.141.0",
    "urllib3==1.26.18",
]
